Create EditQuantityProductDialog
