+++ src/webapp/pages/products/ProductsPage.tsx
@@ -9,11 +9,12 @@ import {
 import React, { ChangeEvent, useCallback, useEffect, useMemo } from "react";
 import i18n from "../../../utils/i18n";
 import SystemUpdateAltIcon from "@material-ui/icons/SystemUpdateAlt";
-import { TextField, Typography } from "@material-ui/core";
+import { Typography } from "@material-ui/core";
 import styled from "styled-components";
 import { useProducts } from "./useProducts";
 import { Product, ProductStatus } from "../../../domain/entities/Product";
-import moment from "moment";
+import { Edit } from "@material-ui/icons";
+import { EditQuantityProductDialog } from "./EditQuantityProductDialog";
 
 export const ProductsPage: React.FC = React.memo(() => {
     const snackbar = useSnackbar();
@@ -97,22 +98,6 @@ export const ProductsPage: React.FC = React.memo(() => {
 
     const tableProps = useObjectsTable(baseConfig, getProducts);
 
-    const handleChangeQuantity = useCallback(
-        (event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) =>
-            onChangeQuantity(event.target.value),
-        [onChangeQuantity]
-    );
-
-    const lastUpdatedCurrentProduct = useMemo(
-        () =>
-            currentProduct
-                ? `${i18n.t("Last updated")}: ${moment(currentProduct.lastUpdated).format(
-                      "yyyy-MM-DD HH:mm"
-                  )}`
-                : undefined,
-        [currentProduct]
-    );
-
     return (
         <Container>
             <Typography variant="h4">{i18n.t("Products")}</Typography>
@@ -122,31 +107,14 @@ export const ProductsPage: React.FC = React.memo(() => {
                 columns={tableProps.columns}
                 onChangeSearch={undefined}
             />
-            {currentProduct !== undefined && (
-                <ConfirmationDialog
-                    isOpen={true}
-                    title={i18n.t("Update Quantity")}
-                    onCancel={cancelEditQuantity}
-                    cancelText={i18n.t("Cancel")}
-                    // eslint-disable-next-line @typescript-eslint/no-misused-promises
-                    onSave={saveEditQuantity}
-                    saveText={i18n.t("Save")}
-                    maxWidth="xs"
-                    fullWidth
-                    disableSave={currentProduct.error !== undefined}
-                >
-                    <Typography style={{ marginBottom: 10 }}>
-                        {lastUpdatedCurrentProduct}
-                    </Typography>
 
-                    <TextField
-                        label={i18n.t("Quantity")}
-                        value={currentProduct.quantity}
-                        onChange={handleChangeQuantity}
-                        error={currentProduct.error !== undefined}
-                        helperText={currentProduct.error}
-                    />
-                </ConfirmationDialog>
+            {currentProduct !== undefined && (
+                <EditQuantityProductDialog
+                    currentProduct={currentProduct}
+                    cancelEditQuantity={cancelEditQuantity}
+                    saveEditQuantity={saveEditQuantity}
+                    onChangeQuantity={onChangeQuantity}
+                />
             )}
         </Container>
     );
